Add Appodeal SDK
WARNING: To achive the integration the gradle config was changed to support multidexing and using the common liberary "com.google.android.gms:play-services-base:17.6.0" between the Firebase SDK and the Appodeal SDK. The native Version of the Appodeal SDK (as IronSource, BidMachine and FacebookAudience-Dependency) was 17.0.0. It seems to work, but could cause trouble in the future. The changelog did just mention Bugfixes and minor changes.